requirements.txt
wrapper-tls-requests
colorama
pyfiglet